feat: complete the src/makefile

+++ src/log.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -7,7 +8,7 @@
 
 /*Log for error level*/
 void log_error(char *s) {
-    int len = strlen(s) + 1;
+    size_t len = strlen(s) + 1;
     char buff[len];
     sprintf(buff, "%s\n", s);
     db_send(buff);    
@@ -15,7 +16,7 @@ void log_error(char *s) {
 
 /*Log for error level with string format.*/
 void log_error_s(char *format, char *s) {
-    int len = strlen(format) + strlen(s);
+    size_t len = strlen(format) + strlen(s);
     char buff[len];
     sprintf(buff, format, s);
     log_error(buff);
@@ -23,7 +24,7 @@ void log_error_s(char *format, char *s) {
 
 /*Log for error level with two strings format*/
 void log_error_s_s(char *format, char *s1, char *s2) {
-    int len = strlen(format) + strlen(s1) + strlen(s2);
+    size_t len = strlen(format) + strlen(s1) + strlen(s2);
     char buff[len];
     sprintf(buff, format, s1, s2);
     log_error(buff);
@@ -31,7 +32,7 @@ void log_error_s_s(char *format, char *s1, char *s2) {
 
 /*Log for error level with a int string format*/
 void log_error_d(char *format, int val) {
-    int len = strlen(format) + 20;
+    size_t len = strlen(format) + 20;
     char buff[len];
     sprintf(buff, format, val);
     log_error(buff);
@@ -39,7 +40,7 @@ void log_error_d(char *format, int val) {
 
 /*Log for error level with a float string format*/
 void log_error_f(char *format, float val) {
-    int len = strlen(format) + 20;
+    size_t len = strlen(format) + 20;
     char buff[len];
     sprintf(buff, format, val);
     log_error(buff);
